// src/types/cms-types.ts
export interface CmsContent<Widgets, Params> {
  active: boolean
  category: null
  categoryId: null
  contentType: string
  defaultLayout: null
  defaultSchema: null
  description: null
  fileUrl: null
  id: ID
  imageUrl: null
  language: string
  layout: null
  metaDescription: null
  metaKeyword: null
  metaTitle: null
  ordering: null
  params: Params
  publicationDate: null
  publicationEndDate: null
  redirect: string
  schema: null
  slug: string
  title: string
  widgetCollectionId: number
  widgets: Widgets
}
// ...
export type Widgets = {
  id: ID
  title: string
  typeId: ID
  collectionId: ID
  point: string
  params: {
    sort_description: {
      value: string
    }
    description: {
      value: string
    }
    image: {
      value: string
    }
    location: {
      value: string
    }
    comment: {
      value: string
    }
    comment_no: {
      value: string
    }
    price: {
      value: string
    }
    discount_price: {
      value: string
    }
    btn_text: {
      value: string
    }
    link: {
      value: string
    }
    tag: {
      value: string
    }
    time: {
      value: string
    }
    view_country: {
      value: string
    }

    destinations: {
      destinations: {
        id: ID
        name: string
        slug: string
        code: string | null
        iata: null
        typeId: ID
        domestic: boolean
      }[]
      value: string
    } | null
  }
  ordering: number
  language: string
  active: boolean
}[]

export type Params = {
  teaser: {
    value: string
  }
  landing_menu: {
    menus: {
      id: ID
      poolId: ID
      parentId: null
      language: string
      title: string
      url: string
      urlTarget: null
      comment: null
      icon: null
      image: null
      fileId: null
      active: boolean
      ordering: number
      createdBy: string
      createdDate: string
      updatedBy: null
      updatedDate: null
      items: {
        id: ID
        poolId: ID
        parentId: null
        language: string
        title: string
        url: string
        urlTarget: null
        comment: null
        icon: null
        image: null
        fileId: null
        active: boolean
        ordering: number
        createdBy: string
        createdDate: string
        updatedBy: null
        updatedDate: null
        items: []
      }[]
    }[]
    value: string
  }
}